LayoutTests/media/modern-media-controls/start-support/start-support-click-to-start-expected.txt

Testing the MediaController click-to-start behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.showsStartButton is true

Pressing on the start button
Media is playing
PASS mediaController.controls.showsStartButton is false

PASS successfullyParsed is true

TEST COMPLETE
